Add buttons and event listeners for one round
Added buttons with js to the body of the DOM with text
for the choices to make (rock, paper, and scissors). Once one is
clicked, on click, run a round of the game. Right now all three
buttons have their own function. Future iteration might have an id
associated with each button and a switch statement that makes the
decision of what to pass. This is event delegation and will improve
performance, although that is likely not an issue right here.