Hard code buttons and use one event listener
Hard coded the buttons and used their ids to have an event
delegation that uses event bubbling to handle the click at the
body level of the code. If I add more to this project, I should
probably put this in a div so other clicks do not trigger
unintended actions.